services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
weaviate-client==3.25.3
Pillow==10.4.0
torch>=2.2.0
torchvision>=0.17.0
transformers==4.44.2
python-multipart==0.0.9
requests==2.32.3

